move convert labels to int to utils
